Day1 complete, starting Day2
2 hours

Day1 issues was that I had to specify [int] in front of the values if I wanted them to add together vs just append two values (1+1 =2 vs 1+1 = 11)

Day2 started. Setup the data import with headers and setting delimiter. Each value has all of its things, so we just need to code the check statement and then increment the counter on success.