Update day2-1.ps1
1 hour

I got the code to work. I had to add quotes around $lettercase to get it to resolve the actual value vs the string. I am not sure why I don't need that everywhere, but maybe it is a double variable thing.

I received the wrong answer since I did not account for the length of the rules to be more than one character long. Wrote some code to start parsing the string, putting values in to a variable until you reach the -, then increment the counter past the For loop to exit on next loop. Not working quite yet, but should.