Day2-1 still
45 min

I had the wrong high/low listed in the inital high function. I also cleared out the count feature when i was quickly trying to implement the changes end of last session. Recieved 346, still incorrect, too low. Previously was 200 something when I wasn't accounting for multi-digit ranges. Adding more values to the Test case and checking the logic to see what it is that i am missing. Looking at a example file, it was supposed to be 483.