3-2 finished, started 4-1
1 1/4 hours

The problem with 3-2 was that I was starting the "count" (What row) with 0. For the pass that took two rows to increment, it was having it out of sync vs assuming the first Row was a match and keying it from there.

For 4-1, It is partially written. It is missing a way to report the last line of the array. I will have it executing the check statements when it get to the carrage return, but there is not one at the end, even if i add on. I also have to write up the code/nested If statement looking for all of the correct values.