day 4-1 done, start 4-2
For 4-1, the process of collection the lines into one was un-necessary. By just searching the line for the required values, i could add them to a variable. Then I checked for a valid passport at that step. I used the carrage return lines to reset the variables. That way the last line missing a carrage return was unnecessary. I added a check to makes sure, if a valid passport was found, it didnt try to find it again before clearing. A line could contain just a cid: which we are ignoring.

for 4-2, I found the -split command that will split a string into an array accoriding to the seperator value that you specify. Currently, I am not getting it to return 4 good passports. Not sure what it is tripping up on yet.