Update day5-2.ps1
1/2 hour
Realized that we could save some steps by attemping to over-write the array values with the translated seat number. That part isn;t working yet. After that, we can sort the list. That also isn;t working. After that, we should be able to go through the list seeing if a number is missing out of the list and looping through the list.