Finish 6-1, started 6-2
1 hour
The Sort Object didn't work as i had it since it was looking as "abc" as a unique entry vs as three characters. Was able to split them into their own entry with .ToCharArray and then sort-object -unique.

I got the last values to process by adding one more sorting process outside of the For loop. Correct answer first try.

6-2 copied in. Trying to first parse the carrage return lines, then check if the Value entry is empty, then do the comparison. Since all of the letters need to exist in all of the grouped results, I should be able to take the first line and use it as a comparitor.